Removed old ViewModel and created new ones
