+++ owncloudApp/src/main/java/com/owncloud/android/presentation/viewmodels/settings/SettingsLogsViewModel.kt
@@ -1,0 +1,45 @@
+/**
+ * ownCloud Android client application
+ *
+ * @author Juan Carlos Garrote Gascón
+ *
+ * Copyright (C) 2021 ownCloud GmbH.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ * <p>
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * <p>
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package com.owncloud.android.presentation.viewmodels.settings
+
+import androidx.lifecycle.ViewModel
+import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
+import com.owncloud.android.presentation.ui.settings.fragments.SettingsLogsFragment
+import com.owncloud.android.providers.LogsProvider
+
+class SettingsLogsViewModel(
+    private val preferencesProvider: SharedPreferencesProvider,
+    private val logsProvider: LogsProvider
+) : ViewModel() {
+
+    fun shouldLogHttpRequests(value: Boolean) = logsProvider.shouldLogHttpRequests(value)
+
+    fun setEnableLogging(value: Boolean) {
+        preferencesProvider.putBoolean(SettingsLogsFragment.PREFERENCE_ENABLE_LOGGING, value)
+        if (value) {
+            logsProvider.startLogging()
+        } else {
+            logsProvider.stopLogging()
+        }
+    }
+
+    fun isLoggingEnabled() = preferencesProvider.getBoolean(SettingsLogsFragment.PREFERENCE_ENABLE_LOGGING, false)
+}
