+++ owncloudApp/src/main/java/com/owncloud/android/presentation/viewmodels/settings/SettingsSecurityViewModel.kt
@@ -22,20 +22,14 @@ package com.owncloud.android.presentation.viewmodels.settings
 
 import android.content.Intent
 import androidx.lifecycle.ViewModel
-import com.owncloud.android.R
 import com.owncloud.android.data.preferences.datasources.SharedPreferencesProvider
 import com.owncloud.android.presentation.UIResult
-import com.owncloud.android.presentation.ui.settings.fragments.SettingsLogsFragment
 import com.owncloud.android.presentation.ui.settings.fragments.SettingsSecurityFragment
-import com.owncloud.android.providers.ContextProvider
-import com.owncloud.android.providers.LogsProvider
 import com.owncloud.android.ui.activity.PassCodeActivity
 import com.owncloud.android.ui.activity.PatternLockActivity
 
-class SettingsViewModel(
-    private val preferencesProvider: SharedPreferencesProvider,
-    private val contextProvider: ContextProvider,
-    private val logsProvider: LogsProvider
+class SettingsSecurityViewModel(
+    private val preferencesProvider: SharedPreferencesProvider
 ) : ViewModel() {
 
     fun isPatternSet() = preferencesProvider.getBoolean(PatternLockActivity.PREFERENCE_SET_PATTERN, false)
@@ -77,35 +71,4 @@ class SettingsViewModel(
 
     fun setPrefTouchesWithOtherVisibleWindows(value: Boolean) =
         preferencesProvider.putBoolean(SettingsSecurityFragment.PREFERENCE_TOUCHES_WITH_OTHER_VISIBLE_WINDOWS, value)
-
-    fun isHelpEnabled() = contextProvider.getBoolean(R.bool.help_enabled)
-
-    fun getHelpUrl() = contextProvider.getString(R.string.url_help)
-
-    fun isSyncEnabled() = contextProvider.getBoolean(R.bool.sync_calendar_contacts_enabled)
-
-    fun getSyncUrl() = contextProvider.getString(R.string.url_sync_calendar_contacts)
-
-    fun isRecommendEnabled() = contextProvider.getBoolean(R.bool.recommend_enabled)
-
-    fun isFeedbackEnabled() = contextProvider.getBoolean(R.bool.feedback_enabled)
-
-    fun isPrivacyPolicyEnabled() = contextProvider.getBoolean(R.bool.privacy_policy_enabled)
-
-    fun isImprintEnabled() = contextProvider.getBoolean(R.bool.imprint_enabled)
-
-    fun getImprintUrl() = contextProvider.getString(R.string.url_imprint)
-
-    fun shouldLogHttpRequests(value: Boolean) = logsProvider.shouldLogHttpRequests(value)
-
-    fun setEnableLogging(value: Boolean) {
-        preferencesProvider.putBoolean(SettingsLogsFragment.PREFERENCE_ENABLE_LOGGING, value)
-        if (value) {
-            logsProvider.startLogging()
-        } else {
-            logsProvider.stopLogging()
-        }
-    }
-
-    fun isLoggingEnabled() = preferencesProvider.getBoolean(SettingsLogsFragment.PREFERENCE_ENABLE_LOGGING, false)
 }
